Claude fixes tests after refactor
