+++ pyproject.toml
@@ -108,3 +108,4 @@ precision = 2
 
 [tool.coverage.html]
 directory = "htmlcov"
+
